WIP: Add a new ResamplerTimer that is attached to the wall clock
This is a new timer that is attached to the wall clock and can be used
to trigger resampling windows. It is a replacement for the channels
Timer class that was attached to the monotonic clock.

This should fix the issue with the resampling windows being too small
when the system clock is not in sync with the wall clock.

Signed-off-by: Leandro Lucarella <luca-frequenz@llucax.com>